Push down errorState into InputAnchor
